u64 for scopeid

+++ benches/basic.rs
@@ -78,7 +78,7 @@ fn app(s: Scope<Root>, n: usize) {
                 count.set(n);
             } else {
                 for i in 0..c {
-                    s.key(i, move |s| {
+                    s.key(i as u32, move |s| {
                         s.button(format!("Item {}", i));
                     });
                 }
